refactor: Move injectables into separate file
